"Add sqrt and pow methods to Field and TryFrom trait implementation for Point
Added sqrt and pow methods to the Field struct and implemented the TryFrom trait for Point struct in the secp256k1 module. These changes were made to provide an easy way to convert from byte array to Point objects and to perform certain mathematical operations on Field objects. The mathematical operations are particularly useful for cryptographic algorithms."