Add README.md with project overview and motivation
A README.md file was added to the project's root directory to give an overview of the project. It details a brief explanation of the implementation of the book "Programming Bitcoin: Learn How to Program Bitcoin from Scratch" in Rust, introduces the book's content, and explains the motivation behind the project. This introduction is meant to help visitors to the codebase quickly understand the nature and motivation of the project.